[pc] Fix power of 2 unit calculation failure on large integers.
It's weird beause we can't use a bigint directly since log(bigint(1<<400))
takes hours (literally). So we convert into a perl scalar so we can use fast
floating point log() (the precision we lose going to a scalar doesn't matter
in the context of log()) and then convert back to a bigint for the pow().